+++ opencti-platform/opencti-front/src/components/dashboard/import-export/DashboardHiddenImportInput.tsx
@@ -1,5 +1,5 @@
-import VisuallyHiddenInput from '../../private/components/common/VisuallyHiddenInput';
-import { useDashboardImport } from './useDashboardImportExport';
+import VisuallyHiddenInput from '../../../private/components/common/VisuallyHiddenInput';
+import useDashboardImport from './useDashboardImport';
 
 const DashboardHiddenImportInput = ({ helpers }: { helpers: ReturnType<typeof useDashboardImport> }) => {
   return <VisuallyHiddenInput type="file" accept="application/JSON" ref={helpers.inputRef} onChange={helpers.onChange} />;
