+++ opencti-platform/opencti-front/src/private/components/workspaces/WorkspaceCreation.tsx
@@ -23,8 +23,8 @@ import { isNotEmptyField } from '../../../utils/utils';
 import Drawer from '../common/drawer/Drawer';
 import { WorkspaceCreationImportMutation } from './__generated__/WorkspaceCreationImportMutation.graphql';
 import { WorkspacesLinesPaginationQuery$variables } from './__generated__/WorkspacesLinesPaginationQuery.graphql';
-import { useDashboardImport } from '../../../components/dashboard/useDashboardImportExport';
-import DashboardHiddenImportInput from '../../../components/dashboard/DashboardHiddenImportInput';
+import useDashboardImport from '../../../components/dashboard/import-export/useDashboardImport';
+import DashboardHiddenImportInput from '../../../components/dashboard/import-export/DashboardHiddenImportInput';
 
 const workspaceMutation = graphql`
   mutation WorkspaceCreationMutation($input: WorkspaceAddInput!) {
